Fix level 5 phpstan lint errors
Note that the version of Plates we are using has an incorrect
docstring for the type passed to registerFunction(), insisting it
is a "callback". However, the Plates docs indicate that the syntax
we are using to register a public method as a template function is
correct.

For now, tell phpstan to ignore the invocations of registerFunction()
until Plates is updated and specifies the correct type for its
second argument in its docstring.