Roll localstack version back
The error logged at https://github.com/localstack/localstack/issues/4973
was occurring in my local setup. This seems to be a problem
with versions of localstack after 0.12.18. Roll back to that version
until a fix is ready for the error. There are no features in later
versions which we are reliant on, so there's no need to keep
using latest tag.