LPAL-651 Improve API exception handling
(This also includes a lot of lint fixes.)

Don't throw runtime errors when a list of applications can't be
retrieved for a user. This causes the template rendering to
break because the AccountInfo view helper doesn't know what to
do with the exception.

This happens when a new user is changing their email/password,
as they don't have any applications yet.

Instead, return an empty list of applications and log API
exceptions. This means that the view can still render correctly,
albeit without LPA applications counts.

This allows the change password/change email to work correctly
when the user's form validation fails (e.g. they type their
new password differently in the two boxes, or they don't repeat
the same email address in the two boxes).