LPAL-651 Prevent raw 500 page when newly-signed-up user changes email/password  (#902)
* LPAL-651 Improve API exception handling

(This also includes a lot of lint fixes.)

Don't throw runtime errors when a list of applications can't be
retrieved for a user. This causes the template rendering to
break because the AccountInfo view helper doesn't know what to
do with the exception.

This happens when a new user is changing their email/password,
as they don't have any applications yet.

Instead, return an empty list of applications and log API
exceptions. This means that the view can still render correctly,
albeit without LPA applications counts.

This allows the change password/change email to work correctly
when the user's form validation fails (e.g. they type their
new password differently in the two boxes, or they don't repeat
the same email address in the two boxes).

* Default to showing verbose S3Monitor logging

* Make it possible to sign in with multiple users

* Add feature to test mismatched passwords

* Add feature to test email change with mismatched emails

* Update non-matching email error messages