Use one package.json for local and docker cypress
The package.json in the root of the project and the dependencies
in the cypress Dockerfile had drifted apart. This meant that running
the cypress tests locally used a different version of cypress and
different plugins and other libraries.

By using the package.json in the root directory of the project,
we ensure that we have equivalent versions of everything when
running cypress ourselves or running it inside the docker container.