LPAL-662 Upgrade laminas-form to v3 (#905)
* Update instructions for modifying composer lock files

This was not updated when I modified the Makefile to streamline
running composer without using docker-compose.

* Update laminas-form to v.3

This also modified a few dependencies.

* Upgrade grunt

* Include vendor folders in docker-compose volume mounts

If these are not included, running the application or tests after
running composer will not mount the new vendor folders, which
is misleading (e.g. unit tests will continue to work as they'll
be using the vendor folder in the image with potentially stale
packages).

* Fix Bootstrap.php so MakeLogger files can be loaded

This is to support running a single phpunit test file from
the command line without having to run the docker container.

Also useful for running individual tests inside phpstorm.

* Update method signature

laminas-form v.3 has different method signature, and PHP 8 is quite
strict, so update the method signature in our subclass to match.

* Checkbox values must be strings, not bools, in laminas-form v.3

* Update package.json in root and fix issues

* Clean up cypress package.json for local use

* Only specify top-level dependencies, not all of their dependencies
* Upgrade to cypress ^8.7.0 (to match Dockerfile)
* Update instructions on using `make cypress-open` and `make cypress-local-shell`
* Remove confusing additional GUI make targets we don't use

* Turn off verbose S3Monitor again

This causes horrible formatting issues when run as part of
the `make cypress-local` target.

* Use one package.json for local and docker cypress

The package.json in the root of the project and the dependencies
in the cypress Dockerfile had drifted apart. This meant that running
the cypress tests locally used a different version of cypress and
different plugins and other libraries.

By using the package.json in the root directory of the project,
we ensure that we have equivalent versions of everything when
running cypress ourselves or running it inside the docker container.

* Fix async vulnerability by updating grunt-contrib-sass

Update to grunt-contrib-sass v2.

This updates other dependencies, but these are only used at
build-time to compile JS and CSS so are a minimal risk.