LPAL-447 Service api psalm lint and upgrade to PHP 8.1 (#913)
* Consistently use PHP 8.0 and composer 2.3

Set the baseline for all docker and composer versions so they match
up.

* Upgrade service-api to PHP 8.1

* Require psalm in service-api composer file

* Update psalm config

* Round 1 of psalm error fixes

* Round 2 of psalm fixes

* Remove references to Laminas\Console\Request

Documentation suggests that onBootstrap() is not called for
CLI commands, so the conditional here is not required.

Also ensure that method signatures for console commands pass
psalm tests and work consistently.

* Round 3 of psalm fixes

* Configure include path for phpunit tests

* Fix unit tests

* Round 4 of psalm fixes

* Allow applicationsService to throw exception

* Table aliases must be passed to Sql->from() method

In PHP 8.1, table aliases cannot be passed directly to the
select() method. Instead, must specify the table and its alias
using from().

* Suppress psalm error about catching laminas-db ExceptionInterface

* Suppress psalm error about null LPA creation date

* Suppress patch() method psalm error

This is due to inconsistencies in laminas-mvc REST controller
methods, where some have a return type of mixed while patch()
has array.

* Update github psalm action config to PHP 8.1 for service-api

* Fix shared logger so service-front unit tests work again

* casper tests need a full-on Linux

* Update to opg-lpa-datamodels 14.0.0

This version supports PHP 8.1.

It also incidentally patches a few other libraries.