Suppress psalm errors for undefined method
We throw a runtime error if the wrong type of adapter is passed
to the constructor, so we know our adapter will have the method
which psalm knows is not on the interface.

This is a terrible pattern but possibly very disruptive to change,
so mitigate by allowing it through psalm and commenting.