Show approximate return by date for LPA if statusDate not available
In the rare case where we don't get a statusDate for an LPA
which is "Return - unpaid", use the less-precise "after 15 working
days" as the predicted date when the user will receive their LPA back.

This should not come into play, but is there in case a regression
is introduced into opg-data-lpa.