Correctly set status on LPAs which have been deleted
If an LPA has been deleted, we get a 410 code back from Sirius.
However, if we previously cached the status of that application
to our db (e.g. when the application was received, we cached a
"Received" stats), the new 410 response didn't overwrite the status
in the cache.

Modify the Application service in service-front so that a 410
code, marked as deleted=true in the API response, correctly
overwrites any cached status with "Waiting".