LPAL-659/LPAL-532 Perfplat containers - security updates + fixes (#932)
* Include snyk in dev deps

* Remove dependency on localstack_client

The localstack_client library did very little and had stopped
working properly, so just duplicate its functionality in our code.

* Fix incorrect instructions in perfplat doc

* Upgrade lambdas and docker images to Python 3.9 where possible

This fixes security issues thrown out by snyk.
Note that this needed a small tweak to a path used in
one of the Dockerfiles, which may be specific to Python 3.9.

* Rationalise perfplatworker Dockerfiles

The separate Dockerfiles for dev and ci are not necessary, and
only have a couple of lines difference. Instead of maintaining
two nearly-identical files, modify docker-compose.yml so it
can use the same Dockerfile as the CI environment.

* Manually update curl to fix security vulnerabilities

* Rationalise and fix vulnerabilities in perfplatapi

* No need to yum update the base image

Only the second image in the Dockerfile is actually distributed,
so there's no need to fix vulnerabilities in the staging "base"
image. This just slows everything down unnecessarily.