LPAL-449 Upgrade service front to PHP 8.1 and lint PHP code (#930)
* Add psalm as a dev dependency

* Add MakeLogger to psalm search path

* Upgrade to PHP 8.1

* Round 1 of psalm level 4 fixes

* Clean out cruft for Sendgrid mail transport

Some redundant classes and tests had been left behind after removal
of the Sendgrid mail transport.

* Temporarily upgrade to dev (v14) branch of opg-lpa-datamodels

* User->name may be null and not a Name object

* Pull in latest patch on opg-lpa-datamodels

* Round 2 of psalm level 4 fixes for service-front

* Clean up API for AuthResponse

* Remove use of static() and just instantiate the class instead
* Move public method into constructor
* Remove tests of internal API

* Improve exception handling in Feedback service

* Catch specific exceptions thrown by the two classes in play
  (ApiClient and mail transport)
* All thrown exceptions should be caught, so @throws is unnecessary

* Suppress psalm errors for undefined method

We throw a runtime error if the wrong type of adapter is passed
to the constructor, so we know our adapter will have the method
which psalm knows is not on the interface.

This is a terrible pattern but possibly very disruptive to change,
so mitigate by allowing it through psalm and commenting.

* Start untangling inconsistent return types from API client

* Consistent handling of return types, response body parsing and exceptions

* Remove unnecessary logging

* Round 3 of psalm level 4 fixes for service-front

(Small one as I just discovered something really horrible and
want a nice clean slate to start from while I address it.)

* Remove code only used in tests

Instead of adding a method specifically for setting the
view, use the public setView() API to do this in the test environment.
Using a ConsoleRenderer causes our helper class under test to
render to a plain string, which we can then assert about in the test.

* Fix method arguments for FormRadio

Instead of passing the attributes to the renderHiddenElement()
method, set them on the element then just pass the element.

(renderHiddenElement() only takes a single argument, so we can't
pass the attributes as the second argument.)

* Simplify ReuseDetails form + element

* Round 4 of psalm level 4 fixes for service-front

* Properly extract referer from request headers

Make sure we cast the retrieved header to Referer so it has
the methods we need to extract the uri from it.

* Round 5 of psalm level 4 fixes for service-front

* Add a convertRequest() method for controllers

This avoids having lots of docblocks to do implicit conversion
and makes it clearer what's going on.

The main reason is to cast the controller request object to
the correct class which has the methods we want to use in our
controllers.

* Round 6 of psalm level 4 fixes for service-front

* Clean up and repair PDF download

Modifications made to the Lpa\Application service had broken
PDF downloads, so repair that.

Also fix download controller and tests.

* Final round of psalm level 4 fixes for service-front

* Update opg-lpa-datamodels (service-front dep) to released version

While doing the upgrade to PHP 8.1, I pointed this at a dev branch.
Now the necessary changes are merged, point to a release tag instead.

* Update psalm github action to PHP 8.1 for service-front

* Update composer dependencies for service-front

* Fix psalm errors in MessageFactory

* Explicitly require laminas-db

psalm fails in github because laminas-db is an optional
composer requirement for laminas-validator. However, we are
definitely using laminas-db, so it's safe to mark that
dependency in composer.json.