Remove README comments for devs outside MoJ
Due to the Notify service requiring a valid key in its constructor,
it's not possible to run the service locally with make dc-up-out
and get something which works.

Remove the notes for developers outside the MoJ until such time
as we are able to provide a way to run the app without integrations.