Fix CVE vulnerabilities
Remove the packages which cause these CVE vulnerabilities:

* CVE-2022-27405 - freetype
* CVE-2022-27406 - freetype

The freetype package is pulled in as a dep of an nginx image
filtering module, but we don't use that module. As it's always
best to remove any packages we're not using anyway, follow advice from
https://github.com/nginxinc/docker-nginx/issues/657#issuecomment-1119372159
and get rid of the offending packages.