Use seeded test user for testing statuses
Note that I had to remove the test for the "Waiting" status. This
is because LPAs in the CI environment may correspond with real
Sirius LPA applications. If we load the dashboard before running
the Statuses feature, those LPAs may get a non-Waiting status
cached for them. If we then attempt to return a "Waiting" response
for those LPAs (i.e. the LPA is not found on Sirius) in our stub,
that response is ignored in favour of the cached response.

This makes it difficult to write a test for "Waiting" which will
pass consistently, as any LPA which gets a Waiting response *but
has not been deleted* is treated as an invalid response; if there
is a cached status, we will then use that status rather than
the invalid response (our assumption being that the LPA either
doesn't exist on Sirius or there was some other kind of error
which prevented us getting its status).