Rejig status view so it can be tested in isolation from the controller
The intention is to directly test the rendered HTML, to replicate
the testing we were doing in cypress, but without having to bring
up the whole stack.

We are of course canning the responses from Sirius (as passed through
our service-api then various transformation stages in service-front
that the data goes through before reaching the controller).

Much of this is covered by the controller tests, but what we're adding
here is a test of the Twig templates themselves, and how it interacts
with the view model, to ensure we have a robust way of detecting any
changes in this area.