Implement simplified LoggerTrait
The MakeLogger LoggerTrait is designed to work with laminas-mvc
apps and will not work correctly in service-admin (which is
not a laminas-mvc app).

Implement a simple alternative logger which can be used within
service-admin as a drop-in replacement for MakeLogger.

As we are not so concerned about logging trace IDs here, we can
resort to logging errors directly and not as JSON. If this is
unacceptable, we will have to consider rewriting the MakeLogger
code so we can output JSON in non-laminas-mvc apps.

The code in this commit improves things, as at the moment I would
expect logging to fail completely for service-admin due to
missing laminas-mvc dependencies.