Passing null to DateTime is deprecated
When we get the last login time for a user, we set this to null
if they have not logged in before. This value is then passed
to new DateTime($lastLogin). However, PHP 8.1 is not happy with
passing a null to the DateTime constructor and throws a deprecation
message. Default this value to 'now' if the user has never logged
in before to fix.