LPAL-800 Sirius gateway - statuses testing (#955)
* Use cypress stubs for call to Sirius gateway

* Remove mocksirius and gateway

* 100% unit test coverage for ProcessingStatus\Service

* Turn on dashboard status tests in CI

As these now only rely on cypress stubs, they can be switched
on in CI.

* Add test for 410 Sirius response code

If an LPA is return unpaid then deleted later, we should see
a Processed status first then a Waiting status (the latter in
response to a 410 status code from Sirius gateway).

* Use seeded test user for testing statuses

Note that I had to remove the test for the "Waiting" status. This
is because LPAs in the CI environment may correspond with real
Sirius LPA applications. If we load the dashboard before running
the Statuses feature, those LPAs may get a non-Waiting status
cached for them. If we then attempt to return a "Waiting" response
for those LPAs (i.e. the LPA is not found on Sirius) in our stub,
that response is ignored in favour of the cached response.

This makes it difficult to write a test for "Waiting" which will
pass consistently, as any LPA which gets a Waiting response *but
has not been deleted* is treated as an invalid response; if there
is a cached status, we will then use that status rather than
the invalid response (our assumption being that the LPA either
doesn't exist on Sirius or there was some other kind of error
which prevented us getting its status).

* Add echo statement to cypress screenshot command

This seems to be the only way to ensure the screenshots
are taken and correctly uploaded. Without this, the screenshots
sometimes just go awol.

* Ensure we wait for Ajax request to finish before checking status lozenges

* Remove cypress helpers for status detail page

Co-authored-by: William Falconer <williamfalconeruk@users.noreply.github.com>