LPAL-840 Prevent 500 errors when user multiple-clicks the delete LPA button (#993)
* Show message to user if they attempt to delete an LPA which is already gone

* Improve JS single use handler

* After link or submit button is clicked, disable further clicks
  by removing the click handler.
* Rationalise the code a little.

* Add cypress feature to test double clicks on delete LPA button

On further testing, I realised the 500 error only occurs on a
triple click. On a double click, it doesn't appear to manifest.

I also discovered that it's not possible to test any more than
a double click in cypress, so we can't automatically test
triple clicking on the delete LPA link.

However, I think this test is still worthwhile, as it ensures that
we don't see an error message or a 500 error when double clicking
on the delete LPA link. It also ensures that the delete LPA button
does actually delete an LPA, which isn't covered elsewhere.

We will have to resort to manual testing for the triple click
scenario, however.

* Run the dashboard LPA delete test in CI