run black on python code
remove logging of password which although in a test, triggers security
checks

LPAL-273 Remove remaining casper tests (#1016)

* Add cypress test for sign out redirect to feedback page

* Remove casperjs functional tests

* Remove casperjs tests from circle build

* Remove references to casper from main Makefile

* Remove unused Makefile

LPAL-842 Improve cypress PDF test (#1015)

* Add pdf-lib as a dev dependency

* Verify that downloaded PDF can be loaded by pdf-lib

moved removal of build-deps, so that xdebug installation now works again

Lpal 878 add 'black'. python formatter to pre-commit hooks, run on all python (#1017)

* add black to pre-commit hook

* run black on python code

* remove logging of password which although in a test, triggers security
checks

Co-authored-by: William Falconer <williamfalconeruk@users.noreply.github.com>

LPAL-877 move certs to region level (#1019)

* Force build

* add dns calculation to region

* move cert defintion to region

LPAL-836 Remove commando library (#998)

* Mount service-api bin directory into local container

* Migrate locking script to Symfony

* Use symfony/console instead of commando

* Don't copy bin folder to service-api docker image

* DynamoDB endpoint is null outside dev

Co-authored-by: William Falconer <williamfalconeruk@users.noreply.github.com>

Use provider-level default tags for account

LPAL-885 Reveal outline around date entry fields (#1025)

* Reveal highlight around date entry fields inside popups

Judicious margin adjustment and overflow change to make the
outline visible around date entry elements.

* Fix outline visibility of date inputs when JS is disabled

When JS is disabled, we also need the overflow to be visible
so that we can see the left-hand edge of the outline on the
day date input.

LPAL-803 Status detail page tests (#987)

* Cleanup of StatusController

* Heading for a refactor of status calculation for detail page

Separate the business logic from the controller action so that
it can be independently tested alongside a Twig renderer.

* Rejig status view so it can be tested in isolation from the controller

The intention is to directly test the rendered HTML, to replicate
the testing we were doing in cypress, but without having to bring
up the whole stack.

We are of course canning the responses from Sirius (as passed through
our service-api then various transformation stages in service-front
that the data goes through before reaching the controller).

Much of this is covered by the controller tests, but what we're adding
here is a test of the Twig templates themselves, and how it interacts
with the view model, to ensure we have a robust way of detecting any
changes in this area.

* Add more test cases

This should now cover all the conditionals in the Twig template
and all the test cases in the original cypress tests.

Co-authored-by: William Falconer <williamfalconeruk@users.noreply.github.com>

LPAL-753 Axe a11y violations fixes  (#1023)

* Remove a11y error caused by duplicate ID

* Remove duplicate ID to clear a11y errors

* Add note about erroneous aria attributes added by govuk_frontend_toolkit

* Don't run a11y checks whenever element is clicked

We were running the OPGCheckA11y() command in cypress every time
we clicked on something. This is going to slow things down a lot
and cause bogus a11y warnings.

For example, because we click on the "Add attorney" button to
add an attorney, and then check a11y straight afterwards, Axe
is faced with a button which is partially hidden under a
transparent overlay. This causes a bogus contrast warning to be
thrown out by Axe, as the button is partially obscured.

We should only run the a11y checks on a page load (i_click_to_visit.js
already caters for this) or after a popup has loaded (if we want to
check popups).

* Reinstate accidentally-deleted cypress step definition

Not really sure how the cypress tests carried on working
without this step...

* Correspondent added via API must match seeded data

For the cypress features to work in the cypress GUI and in CI,
the correspondent added through the API must have the same
name and address as the seeded correspondent.

* Only add an id to the first child div of a details element

Tighten up the selector on <details> elements so that only
the first child div is assigned a content ID.

(This is used to polyfill the details element on browsers where
it has no native implementation. By convention, we use a single
<div> inside the details element to hold its expandable content.
But the selector previously used would assign an ID to all of the
child divs of the details element, which caused duplicate id
attributes on child elements.)

* Do a11y checks on transitioning to new page

Rather than checking a11y compliance on clicks (as we were
previously), check when we transition to a new page ("I am taken
to..." etc.). This means we check a11y at the same time as we
check whether we've moved to the right page.

Removing the a11y check on click appears to have removed a lot
of the a11y checks, so putting them back in here should help
highlight issues again.