Rename donor checkbox to cannotSign
The donor popup checkbox was called canSign, but needed to
be checked to indicate that the donor *cannot* sign. Change the
name of the form field to reflect its proper purpose.

However, the field in the database is called canSign. This therefore
needs to store the inverse of the cannot sign form checkbox: if
the box is checked, canSign=false; if it is not checked, canSign=true.
This requires some gymnastics in the form to invert the cannotSign
value into the canSign value.