Move SignupAndChangeDetails to 'other cypress tests'
Don't run this test under the @Signup tag, as it adds a minute
to every test step. This is actually a separate signup test
which can run independently of the main signup tests, and only
needs to run once. It can be moved to the other tests as a result.