Don't run a11y checks whenever element is clicked
We were running the OPGCheckA11y() command in cypress every time
we clicked on something. This is going to slow things down a lot
and cause bogus a11y warnings.

For example, because we click on the "Add attorney" button to
add an attorney, and then check a11y straight afterwards, Axe
is faced with a button which is partially hidden under a
transparent overlay. This causes a bogus contrast warning to be
thrown out by Axe, as the button is partially obscured.

We should only run the a11y checks on a page load (i_click_to_visit.js
already caters for this) or after a popup has loaded (if we want to
check popups).