Do a11y checks on transitioning to new page
Rather than checking a11y compliance on clicks (as we were
previously), check when we transition to a new page ("I am taken
to..." etc.). This means we check a11y at the same time as we
check whether we've moved to the right page.

Removing the a11y check on click appears to have removed a lot
of the a11y checks, so putting them back in here should help
highlight issues again.