LPAL-753 Axe a11y violations fixes  (#1023)
* Remove a11y error caused by duplicate ID

* Remove duplicate ID to clear a11y errors

* Add note about erroneous aria attributes added by govuk_frontend_toolkit

* Don't run a11y checks whenever element is clicked

We were running the OPGCheckA11y() command in cypress every time
we clicked on something. This is going to slow things down a lot
and cause bogus a11y warnings.

For example, because we click on the "Add attorney" button to
add an attorney, and then check a11y straight afterwards, Axe
is faced with a button which is partially hidden under a
transparent overlay. This causes a bogus contrast warning to be
thrown out by Axe, as the button is partially obscured.

We should only run the a11y checks on a page load (i_click_to_visit.js
already caters for this) or after a popup has loaded (if we want to
check popups).

* Reinstate accidentally-deleted cypress step definition

Not really sure how the cypress tests carried on working
without this step...

* Correspondent added via API must match seeded data

For the cypress features to work in the cypress GUI and in CI,
the correspondent added through the API must have the same
name and address as the seeded correspondent.

* Only add an id to the first child div of a details element

Tighten up the selector on <details> elements so that only
the first child div is assigned a content ID.

(This is used to polyfill the details element on browsers where
it has no native implementation. By convention, we use a single
<div> inside the details element to hold its expandable content.
But the selector previously used would assign an ID to all of the
child divs of the details element, which caused duplicate id
attributes on child elements.)

* Do a11y checks on transitioning to new page

Rather than checking a11y compliance on clicks (as we were
previously), check when we transition to a new page ("I am taken
to..." etc.). This means we check a11y at the same time as we
check whether we've moved to the right page.

Removing the a11y check on click appears to have removed a lot
of the a11y checks, so putting them back in here should help
highlight issues again.