Reorganise writing PDF files to disk
Intermediate PDF files are removed from disk as soon as the PDF
object is done with. This makes refactoring difficult, as you
lose the PDF file as soon as its object is garbage collected.

To manage this, modify the PDF renderer so that it returns both
the filepath of the PDF and its content (as a binary string).
This can then be written to the local filesystem (for generating
PDFs for testing purposes) or to S3.

The intermediate PDFs are still garbage-collected, and in the
online worker will be removed as soon as they are put to S3.
However, we can manually write the PDF content to the local disk
for testing as we don't lose the PDF content immediately.