Scale size and offsets for PNG barcode
Generate a PNG barcode which is A4-sized, modifying the offsets
and factor (which scales the barcode bar widths and height) so that
the barcode is in the correct position. This will still need
to be converted to a PDF anyway, as Pdftk can't cope with using
a PNG as a stamp file and fails silently.