LPAL-862 Fix donor popup "able to sign" checkbox (#1018)
* Rename donor checkbox to cannotSign

The donor popup checkbox was called canSign, but needed to
be checked to indicate that the donor *cannot* sign. Change the
name of the form field to reflect its proper purpose.

However, the field in the database is called canSign. This therefore
needs to store the inverse of the cannot sign form checkbox: if
the box is checked, canSign=false; if it is not checked, canSign=true.
This requires some gymnastics in the form to invert the cannotSign
value into the canSign value.

* Checkboxes must be visible to be checked

The cypress test for donor popup failed because the "cannot sign"
checkbox was not in the viewport. Ensure it is scrolled into
view before attempting to check it.

Also add an additional assertion to ensure that the checkbox
remains checked when the popup is opened a second time.

* Move SignupAndChangeDetails to 'other cypress tests'

Don't run this test under the @Signup tag, as it adds a minute
to every test step. This is actually a separate signup test
which can run independently of the main signup tests, and only
needs to run once. It can be moved to the other tests as a result.

Co-authored-by: William Falconer <williamfalconeruk@users.noreply.github.com>