Use db cache for status if Sirius opg-data-lpa unavailable
WWe are already fetching the LPA from the database, so poke into
its metadata (cached in the db) for the status if we couldn't
get it live.

This is hard to write a cypress test for, as you'll only see
it when the mocksirius is unavailable locally. To test, run
the cypress suite once with the full stack up; then
stop the gateway container then run the cypress tests again.
If they pass the second time, it means the cached metadata
is being used to populate the view status message page.