Default to leaving the 'cannot sign' checkbox unchecked
The value of the canSign field in the db is not provided to
the donor form when reusing details. This means that we can't
rely on it when setting the checkbox for a reused donor, and
that the existing logic results in a checked "cannot sign"
checkbox whenever a donor is reused.

Change this behaviour so that a reused donor always results in
an unchecked "cannot sign" checkbox. This returns the
donor form to the behaviour it had prior to the fixes put in
for LPAL-862.

An enhancement to get the canSign value for the donor will be
added via a separate ticket.