Add cypress tag for tests which include their own signup steps
Any cypress test which should not be run as part of the main "remaining
tests" group can be tagged with @SignupIncluded. Any tests so
tagged run in a separate cypress container.

Also modify start.sh so that, if SignupIncluded is in the CYPRESS_TAGS
env var, we don't run the @Signup feature as part of start.sh.

These tests should include their own signup sequence and not
rely on any pre-seeded test users.