Correctly ignore features which include their own signup steps
The previous regex caused tests which should include signup
to not perform the signup step in start.sh.

Change to an @IgnoreSignup tag on features, and only include
features in the "remaining tests" run *without* this tag.

The "signup included" features with this tag can then be run
as a separate group.