Improve error handling around PDF retrieval
If the HTTP GET to the API service fails, the PDF info retrieved
is a bool, rather than an array. This then causes a 500 error
when the app tries to retrieve a key from the bool.

Clean that up so we make sure we have an array before accessing
it. Also put in some additional logging when errors occur.