LPAL-907/LPAL-908 Set "cannot sign" for donor to value in reused LPA (#1038)
* Default to leaving the 'cannot sign' checkbox unchecked

The value of the canSign field in the db is not provided to
the donor form when reusing details. This means that we can't
rely on it when setting the checkbox for a reused donor, and
that the existing logic results in a checked "cannot sign"
checkbox whenever a donor is reused.

Change this behaviour so that a reused donor always results in
an unchecked "cannot sign" checkbox. This returns the
donor form to the behaviour it had prior to the fixes put in
for LPAL-862.

An enhancement to get the canSign value for the donor will be
added via a separate ticket.

* Include canSign field for donor reuse

The canSign field is filtered out by default for all actors.
If this field is available, retain it so that we can fill
the "cannot sign" checkbox in the donor form when reusing a
donor from a previous LPA application.

* Add cypress test to check reused donor retains "can't sign" value

* Add reuse test to circleci

Also rework the tests slightly so that the two tests which
are separated out are excluded from the remaining tests. It's
likely these are part of the reason the build times have increased.

* Add cypress tag for tests which include their own signup steps

Any cypress test which should not be run as part of the main "remaining
tests" group can be tagged with @SignupIncluded. Any tests so
tagged run in a separate cypress container.

Also modify start.sh so that, if SignupIncluded is in the CYPRESS_TAGS
env var, we don't run the @Signup feature as part of start.sh.

These tests should include their own signup sequence and not
rely on any pre-seeded test users.

* Correctly ignore features which include their own signup steps

The previous regex caused tests which should include signup
to not perform the signup step in start.sh.

Change to an @IgnoreSignup tag on features, and only include
features in the "remaining tests" run *without* this tag.

The "signup included" features with this tag can then be run
as a separate group.

* Run signup as part of local cypress tests

Co-authored-by: William Falconer <williamfalconeruk@users.noreply.github.com>