LPAL-911 Improve cypress scripts (#1052)
* First pass at unifying cypress scripts

* Remove userNumber.js (now done in Python)
* Replicate functionality of start.sh in Python
* Move S3Monitor into a class
* Start S3Monitor on a thread in main script

Arguments are not pulled from command line yet, but hard-coded.

Also need to get the return code from cypress and use that
as return code for whole script.

* Pass environment variables correctly

cypress doesn't recognise some environment variables if they
are passed with the -e flag, so set those as environment variables
set on the command line before cypress is invoked.

Also capture return code from cypress and use it as the exit
code for the script.

* Introduce "runs": groups of tags to run in sequence

By passing groups of tags separated by commas, e.g.

@Signup,@StitchedPF or @StitchedHW

the script will split around the comma and then run features
tagged with @Signup, followed by features tagged with
@StitchedPF or @StitchedHW, in sequence. This mimics the
existing functionality but allows any sequence of tags
to be specified, making it simpler to exclude signup where it's
not required, and making it more explicit where signup is
being run before other features.

* Enable running sequences of tags in groups and reimplement stitch

* Use python cypress runner in Makefile

* Fallback to env vars if CLI flags not set

* Add another env var used in cypress test helpers

* Don't run @Signup twice

Because the shared signup tests are gathered using negative
filters (i.e. "run everything *except* these tags"), we have to
be explicit about which tests we shouldn't run. In this case,
we don't want to run the @Signup feature again, as we explicitly
run it as the first step.

* Fix output when script can't run

* Use return code from python as exit code for bash